create custom_attribute_widgets file
